Add files via upload
Exercise 1.2 Launch Spyder.
Type in the program above (don’t worry yet about the details), save it as a Python file (text
file with a .py extension), run it and check that it works. (The button with the green triangle
runs the current Python file, or else you can type run filename.py in the IPython command
line.) 

There was an image from which I copied the code in this file.